Safari-609.1.11.1/LayoutTests/http/tests/security/dataURL/xss-DENIED-to-data-url-sub-frame-expected.txt

CONSOLE MESSAGE: line 194: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame. Protocols, domains, and ports must match.
This tests that the main frame can't access the contents of an iframe that contains a data: URL loaded page


PASS: Cross frame access to a data: URL was denied.


--------
Frame: 'aFrame'
--------
Inner iframe.
